Revision 170 Bringup
I initially decided not to use the current progress of RuneLite as of Revision 170. I have changed my mind in the hopes that refactoring will continue. I will keep this updated at revision releases. Please enjoy the work that Ada and his contributors worked very hard to create. Be sure to give them credit if you use this in your projects.