rm RL Reflection
quite a curious bit of code, seems to be nothing more than a middleman.

Refactored to conform to jagex standard code.

Removed RuneLite api as it is no longer necessary.

Should also stop the weird in game errors dealing with the packetBuffer.

Added Launcher for testing, will persist   in future updates.